refactor: modules.js 파일 내 함수 리팩토링 reactiveItemCount 함수 삭제 후 Items.jsx compornent 내 구현
